Reorganize and small update
