+++ app/src/main/java/com/peter/georeminder/MainScreen.java
@@ -14,14 +14,12 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.Vibrator;
 import android.preference.PreferenceManager;
-import android.preference.PreferenceScreen;
 import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v4.app.ActivityOptionsCompat;
 import android.support.v4.content.ContextCompat;
-import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -56,7 +54,7 @@ import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem;
 import com.mikepenz.materialdrawer.model.interfaces.IProfile;
 import com.parse.ParseObject;
 import com.peter.georeminder.models.Reminder;
-import com.peter.georeminder.utils.RecyclerAdapter;
+import com.peter.georeminder.utils.recyclerview.RecyclerAdapter;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -695,6 +693,8 @@ public class MainScreen extends AppCompatActivity{
                             @Override
                             public boolean onItemClick(View view, int position, IDrawerItem drawerItem) {
                                 Intent toLoginScreen = new Intent(MainScreen.this, LoginScreen.class);
+                                if(Build.VERSION.SDK_INT >= 21)
+                                    getWindow().setExitTransition(null);
                                 startActivityForResult(toLoginScreen, LOGIN_REQUEST_CODE, ActivityOptionsCompat.makeSceneTransitionAnimation(MainScreen.this).toBundle());
                                 Handler handler = new Handler();
                                 handler.postDelayed(new Runnable() {
