+++ app/src/main/java/com/peter/georeminder/utils/login/ProgressGenerator.java
@@ -1,4 +1,4 @@
-package com.peter.georeminder.utils;
+package com.peter.georeminder.utils.login;
 
 
 //        ("`-''-/").___..--''"`-._
