+++ app/src/main/res/layout/activity_google_map_screen.xml
@@ -6,10 +6,25 @@
     tools:context="com.peter.georeminder.WholeMapScreen"
     android:id="@+id/google_map_drawer_layout">
 
-    <include
-        layout="@layout/content_google_map_screen"
+    <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent" >
+
+        <fragment
+            android:id="@+id/whole_google_map"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:name="com.google.android.gms.maps.SupportMapFragment"/>
+
+        <com.quinny898.library.persistentsearch.SearchBox
+            android:id="@+id/google_map_searchBox"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentTop="true"
+            android:layout_marginTop="25dp"/>
+
+    </RelativeLayout>
 
     <android.support.design.widget.NavigationView
         android:id="@+id/nav_google_map_view"
