+++ app/src/main/res/layout/activity_main_screen.xml
@@ -58,7 +58,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="@dimen/fab_margin"
-        android:layout_gravity="bottom|right"
+        android:layout_gravity="bottom|end"
         app:menu_icon="@drawable/fab_add"
         app:menu_colorNormal="@color/colorPrimary"
         app:menu_labels_showAnimation="@anim/fab_slide_in_from_right"
