+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -12,6 +12,7 @@
         android:src="@mipmap/ic_launcher"
         android:layout_marginTop="80dp"
         android:layout_centerHorizontal="true"
-        android:visibility="invisible"/>
+        android:visibility="invisible"
+        android:contentDescription="@string/img_logo" />
 
 </RelativeLayout>
