--- app/src/main/res/layout/content_amap_map_screen.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:map="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.peter.georeminder.WholeMapScreen"><!--TODO: might need to change-->
-
-    <com.amap.api.maps.MapView
-        android:id="@+id/whole_amap_map"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-
-    <com.quinny898.library.persistentsearch.SearchBox
-        android:id="@+id/amap_map_searchBox"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginTop="25dp"/>
-
-</RelativeLayout>
