+++ app/src/main/res/layout/nav_header_whole_map.xml
@@ -21,7 +21,8 @@
         android:layout_above="@+id/nav_head_appname"
         android:src="@mipmap/ic_default_avatar"
         android:background="@null"
-        style="?android:attr/borderlessButtonStyle"/>
+        style="?android:attr/borderlessButtonStyle"
+        android:contentDescription="@string/img_avatar"/>
 
     <TextView
         android:id="@+id/nav_head_appname"
@@ -41,8 +42,8 @@
         android:layout_height="wrap_content"
         android:text="@string/nav_head_username"
         android:textColor="@color/colorPrimaryDark"
-        android:textSize="20dp"
+        android:textSize="20sp"
         android:layout_alignParentBottom="true"
-        android:layout_alignParentLeft="true"/>
+        android:layout_alignParentStart="true"/>
 
 </RelativeLayout>
