+++ app/src/main/res/layout/reminder_geo_edit_screen.xml
@@ -25,7 +25,7 @@
 
         <TextView
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
+            android:layout_height="wrap_content"
             android:text="@string/long_text"/>
 
     </ScrollView>
