+++ app/src/main/res/layout/reminder_normal_edit_screen.xml
@@ -6,7 +6,7 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:textSize="54dp"
+        android:textSize="54sp"
         android:text="Normal"/>
 
 </LinearLayout>
