+++ app/src/main/res/layout/reminder_recycler_item.xml
@@ -27,14 +27,15 @@
                 android:layout_height="120dp"
                 android:layout_margin="10dp"
                 android:layout_centerVertical="true"
-                android:layout_alignParentLeft="true"
-                android:src="@mipmap/ic_launcher"/>
+                android:layout_alignParentStart="true"
+                android:src="@mipmap/ic_launcher"
+                android:contentDescription="@string/img_map"/>
 
             <RelativeLayout
                 android:id="@+id/content_relative_layout"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:layout_toRightOf="@+id/recycler_item_map_screenshot"
+                android:layout_toEndOf="@+id/recycler_item_map_screenshot"
                 android:layout_centerVertical="true">
 
                 <TextView
@@ -44,7 +45,7 @@
                     android:layout_centerHorizontal="true"
                     android:text="Title"
                     android:ellipsize="end"
-                    android:textSize="24dp"
+                    android:textSize="24sp"
                     android:layout_marginTop="20dp"/>
 
                 <TextView
@@ -56,7 +57,7 @@
                     android:layout_marginBottom="20dp"
                     android:text="This is the content for an geo reminder"
                     android:ellipsize="marquee"
-                    android:textSize="12dp"
+                    android:textSize="12sp"
                     />
 
             </RelativeLayout>
